Docker build use .aws-architecture
Why these changes are being introduced:

When a .aws-architecture file is present, docker builds should
use this architecture.  This will come organically when we get
into AWS ECR builds, but this application also includes a
convenience 'make docker-build' that should utilize it too.

How this addresses that need:
* Adds 'CPU_ARCH' env var to top of Makefile

Side effects of this change:
* None

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/USE-111