Stub base model class and CLI download-model command
Why these changes are being introduced:

For this application to create embeddings, it will need some
structure to handle the downloading, loading, and use of
particular models.

Additionally, this application should provide the opinionated
functionality to download and zip up a model, either for use
within a Docker image, or local testing.

How this addresses that need:
* Class 'BaseEmbeddingModel' is created.  This base class will be
extended by actual models we intend to use.  This base class will
expect any child class to define methods for downloading, loading,
and using the model.

* Stubs a CLI command 'download-model' which we will use very
soon for building a Docker image that contains the model weights.

Side effects of this change:
* None at this time.

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/USE-112
* https://mitlibraries.atlassian.net/browse/USE-113