Rework model registration and loading
Why these changes are being introduced:

Previously, a model's URI was only codified in the registry.py file, and not
directly obvious from a class itself aside from docstrings.

How this addresses that need:

Reworks the base class to require a MODEL_URI attribute.  This attribute
is used to generate a model_uri:class dictionary that is used for registration
of models.

Side effects of this change:
* None

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/USE-113