Download and configure model on Docker image build
Why these changes are being introduced:

We have opted to include the model weights and assets inside the Docker image.  When the model is
small-ish like ours, this avoids the need to save the model to S3 and download each time the CLI is
invoked.

How this addresses that need:

The CLI command `download-model` is used within the Dockerfile itself to download the model.  As noted
via inline comments, the env vars `TE_MODEL_URI` and `TE_MODEL_DOWNLOAD_PATH` are also set in the Dockerfile.
This has a dual purpose.  First, these env vars inform the `download-model` CLI invocation within the
Dockerfile.  Second, they persist to the container and establish the model as default for all calls.

Side effects of this change:
* On Docker image build, the model will be downloaded from HuggingFace, configured, and included
in the final Docker image.

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/USE-113