Lean into model_required CLI decorator
Why these changes are being introduced:

Many of the  CLI commands will require an embedding class and model to work.
A decorator was created originally that injected a --model-uri CLI argument,
but it also provides a place to load the class itself and become more of a
middleware.

How this addresses that need:
Updates the model_required decorator to also load the embedding model class.
This DRY's up the CLI commands that use it and centralizes that logic and
conventions for the CLI argument, env vars, and whatnot.

Lastly, it is now required to include a 'model_path' when instantiating
a model class instance, and this location is used for both download and
load.

Side effects of this change:
* None

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/USE-112