Rename RecordText to EmbeddingInput
Why these changes are being introduced:

Code review suggested that 'RecordText' was a confusing name
for the object that we prepare to then create an embedding from.

How this addresses that need:

Renamign to 'EmbeddingInput' makes it crystal clear that we are
preparing an object that will be used to create an embedding.

Side effects of this change:
* None

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/USE-112