Store sparse vector and decoded token weights
Why these changes are being introduced:

Formerly, our 'Embedding' class only had an 'embedding' property for the output.
However, for our first model in the pipeline, opensearch-project/
opensearch-neural-sparse-encoding-doc-v3-gte, it produces two representations of the
embedding that are useful to store: a sparse vector and decoded token weights.

How this addresses that need:

Updates the 'Embedding' class to explicitly store both representations of the
embedding.

We may decide that we don't store both, or some futures models may not produce
decoded token weights of any kind, but this matches our first proposed model
and pipeline.  Better to be explicit and opinionated in these early days, then
adjust later if needed.

Side effects of this change:
* None

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/USE-112