Introduce strategies framework for preparing EmbeddingInputs
Why these changes are being introduced:

A core requirement of this application is the ability to take a TIMDEX
JSON record and "transform" all or parts of it into a single string for
which an embedding can be created.  We are calling these "embedding
strategies" in the context of this app.

While our first strategy will likely be a very simple, full record approach,
we want to support multiple strategies in the application, and even
multiple strategies for a single record in a single invocation.

How this addresses that need:
* A new 'strategies' module is created
* A base 'BaseStrategy' class, with a required 'extract_text()' method
for implementations
* Our first strategy represented in class 'FullRecordStrategy', which
JSON dumps the entire TIMDEX JSON record.
* A registry of strategies, similar to our models, that allow CLI
level validation.

Side effects of this change:
* None really, but further solidifies that this application is contains
the opinionation about how text is prepared for the embedding process.

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/USE-131
* https://mitlibraries.atlassian.net/browse/USE-132