OSNeuralSparseDocV3GTE embedding business logic
Why these changes are being introduced:

Our first embedding class OSNeuralSparseDocV3GTE was ready for a real
create_embedding() method with the rest of the moving parts mostly complete at this time.

How this addresses that need:

The HuggingFace model card, https://huggingface.co/
opensearch-project/opensearch-neural-sparse-encoding-doc-v3-gte, contains a section of
example code for using this model with the transformers library. This logic was ported
over to our class, with the model already downloaded and loaded taken care of.

The biggest addition is the method _decode_sparse_vectors() which converts the numerical
sparse vector into a dictionary of token:weights.  This decoded token weight form is what
we'll pass directly to OpenSearch.

With the new functionality in place, the tests associated with this class were also
updated.  Fixtures were moved into the testing file, a pattern we could adopt for any
future models to keep them out of the shared conftest.py.

Side effects of this change:
* CLI is capable of producing embeddings for our first model OSNeuralSparseDocV3GTE

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/USE-136