Switch to sentence-transformers from transformers
Why these changes are being introduced:

The first approach for the embedding class OSNeuralSparseDocV3GTE used
the `transformers` library for creating embeddings.  Some early exploratory
code seemed to indicate this more low level library would provide more flexibility
in response formats and better performance.

However, when exploring using multiprocessing for the `transformers` library,
the alternate approach of using the `sentence-transformers` library was explored
given it's more out-of-the-box multiprocessing support.  During that exploration,
based on learnings since the original spike code, it was determined that the
`sentence-transformers` library might be a better fit overall for our purposes.

Pivoting to this library will simplify our actual embedding logic, while providing
some out-of-the-box tuning capabilities that should be sufficient for our
purposes.

How this addresses that need:

The embedding class `OSNeuralSparseDocV3GTE` is reworked to use
`sentence-transformers` instead of `transformers` for creating embeddings.

This reducues considerably complexity in the actual creating of embeddings,
while also exposing an API for multiprocessing.  It's worth noting that testing
is indicating that multiprocessing will *not* speed up embeddings, at least for
the contexts we aim to create them, but the `sentence-transformers` library also
better handles parallelism without explicit multiprocessing.

In summary, switching to `sentence-transformers` results in a simpler API
for creating embeddings, better out-of-the-box performance, with an API
that still allows for more tuning later.

Side effects of this change:
* None

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/USE-137