Update full_record strategy
Why these changes are being introduced:

With the json.dumps() approach, we had a lot of JSON characters and cruft in the
final string used for embedding.

How this addresses that need:

Produces a <field>:<value><newline> repeating structure from the full record.

Side effects of this change:
* None

Relevant ticket(s):
* None