Do not save sparse vectors for OSNeuralSparseDocV3GTE
Why these changes are being introduced:

Our initial pass with the embedding class OSNeuralSparseDocV3GTE was to save both
the sparse vector and the decoded token:weights.  Each sparse vector was the length of the
model vocabulary, about 30k, with mostly zeros.  While technically this could be used for
analysis beyond just the decoded token:weights given to OpenSearch, the data transfer
and storage overhead exceeds any known use cases at the moment.

How this addresses that need:

The OSNeuralSparseDocV3GTE embedding model is updated to not include the sparse vector
for the Embedding.embedding_vector property on output.

This can easily be turned on later, with an inline code comment showing how to toggle
that back on.

Side effects of this change:
* No sparse vectors are stored for now, storage is decreased.

Relevant ticket(s):
* None